Create data namespace
