+++ cartoframes/data/__init__.py
@@ -1,0 +1,8 @@
+from __future__ import absolute_import
+
+from .dataset import Dataset, get_query, get_geodataframe
+from .dataset_info import DatasetInfo
+
+__all__ = [
+    'Dataset'
+]
