+++ cartoframes/data/utils.py
@@ -1,4 +1,10 @@
+import time
 import binascii as ba
+
+from warnings import warn
+from carto.exceptions import CartoException, CartoRateLimitException
+
+from ..columns import Column
 
 try:
     import geopandas
@@ -6,6 +12,8 @@ try:
 except ImportError:
     HAS_GEOPANDAS = False
 
+
+DEFAULT_RETRY_TIMES = 3
 
 GEOM_COLUMN_NAMES = [
     'geometry',
@@ -113,3 +121,30 @@ def decode_geometry(ewkb):
                             return wkt.loads(ewkb)
                         except Exception:
                             pass
+
+
+def recursive_read(context, query, retry_times=DEFAULT_RETRY_TIMES):
+    try:
+        return context.copy_client.copyto_stream(query)
+    except CartoRateLimitException as err:
+        if retry_times > 0:
+            retry_times -= 1
+            warn('Read call rate limited. Waiting {s} seconds'.format(s=err.retry_after))
+            time.sleep(err.retry_after)
+            warn('Retrying...')
+            return recursive_read(context, query, retry_times=retry_times)
+        else:
+            warn(('Read call was rate-limited. '
+                  'This usually happens when there are multiple queries being read at the same time.'))
+            raise err
+
+
+def get_columns(context, query):
+    col_query = '''SELECT * FROM ({query}) _q LIMIT 0'''.format(query=query)
+    table_info = context.sql_client.send(col_query)
+    return Column.from_sql_api_fields(table_info['fields'])
+
+
+def setting_value_exception(prop, value):
+    return CartoException(("Error setting {prop}. You must use the `update` method: "
+                           "dataset_info.update({prop}='{value}')").format(prop=prop, value=value))
