--- cartoframes/datasets/__init__.py
@@ -1,7 +1,0 @@
-from .dataset import Dataset, recursive_read, _decode_geom, get_columns, get_query
-from .dataset_info import DatasetInfo, setting_value_exception
-
-__all__ = [
-    'Dataset',
-    'DatasetInfo'
-]
